refactor: update api service
