+++ api/contact.ts
@@ -1,7 +1,23 @@
-import Firestore from '~/plugins/firebase'
+import Firestore from '../plugins/firebase'
 import dayjs from 'dayjs'
 
-import { ContactCategories } from '~/services/contact'
+export const ContactCategories: Array<{
+  value: number
+  text: string
+}> = [
+  {
+    value: 1,
+    text: '仕事のご依頼'
+  },
+  {
+    value: 2,
+    text: '当ブログへのご提案'
+  },
+  {
+    value: 99,
+    text: 'その他'
+  }
+]
 
 type ContactCategory = {
   value: number
