--- services/blog.ts
@@ -1,5 +1,0 @@
-export const PAGE = 10
-
-export const LATEST_PAGE = 6
-
-export const ORDER = '-fields.publishDate'
