--- services/contact.ts
@@ -1,17 +1,0 @@
-export const ContactCategories: Array<{
-  value: number
-  text: string
-}> = [
-  {
-    value: 1,
-    text: '仕事のご依頼'
-  },
-  {
-    value: 2,
-    text: '当ブログへのご提案'
-  },
-  {
-    value: 99,
-    text: 'その他'
-  }
-]
