--- services/release.ts
@@ -1,1 +1,0 @@
-export const blogVersion: number = 2.0
